Change to Moose. Kill config accessor
All the config may as well (and should be) stored directly in the object instance. ->_config is a throwback used in the modules which _ALSO_ can be an
old style plugin composed onto the app class. For a new module like this with no back compat needed, it's totally superfluous